pyproject.toml
[project]
name = "langchain-nayak"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain>=1.2.9",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.8",
    "python-dotenv>=1.2.1",
]
